Change artifactId for flies war to match directory name 'flies-war'.
This is to avoid breaking URLs, eg when we use 'mvn site'.

"An Important Note from Brian Fox: Naming the folder after the artifactId is
critical if you expect inherited urls (scm/site) to work. It always appends
the artifactId to the url as it?s inherited. And for the same reason, you
want them to match what?s in the scm."
http://www.sonatype.com/people/2011/01/maven-tip-project-directories-and-artifact-ids/